+++ src/main/java/com/ll/domain/member/controller/ApiV1MemberController.java
@@ -1,0 +1,9 @@
+package com.ll.domain.member.controller;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+public class ApiV1MemberController {
+}
