+++ src/main/java/com/ll/domain/member/repository/MemberRepository.java
@@ -1,0 +1,4 @@
+package com.ll.domain.member.repository;
+
+public interface MemberRepository {
+}
