+++ src/main/java/com/ll/domain/member/service/MemberService.java
@@ -1,0 +1,9 @@
+package com.ll.domain.member.service;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class MemberService {
+}
