+++ src/main/java/com/ll/domain/order/controller/ApiV1OrderController.java
@@ -1,0 +1,9 @@
+package com.ll.domain.order.controller;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+public class ApiV1OrderController {
+}
