+++ src/main/java/com/ll/domain/order/repository/OrderRepository.java
@@ -1,0 +1,4 @@
+package com.ll.domain.order.repository;
+
+public interface OrderRepository {
+}
