+++ src/main/java/com/ll/domain/order/repository/WishListRepository.java
@@ -1,0 +1,4 @@
+package com.ll.domain.order.repository;
+
+public interface WishListRepository {
+}
