+++ src/main/java/com/ll/domain/order/service/OrderService.java
@@ -1,0 +1,9 @@
+package com.ll.domain.order.service;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class OrderService {
+}
