+++ src/main/java/com/ll/domain/product/controller/ApiV1ProductController.java
@@ -1,0 +1,9 @@
+package com.ll.domain.product.controller;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+public class ApiV1ProductController {
+}
