+++ src/main/java/com/ll/domain/product/repository/ProductRepository.java
@@ -1,0 +1,4 @@
+package com.ll.domain.product.repository;
+
+public interface ProductRepository {
+}
