+++ src/main/java/com/ll/domain/product/service/ProductService.java
@@ -1,0 +1,9 @@
+package com.ll.domain.product.service;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class ProductService {
+}
